Remove Immutable.
While enforcing state immutability is good in theory, in practice
we've never had any problems with it - it's simple enough to just
learn to use object spread operator wherever necessary. In our
opinion, the lack of ability to use standard JS syntext when interacting
with objects is not worth it.